compare-timetables: extract gradientUtils
