--- static/js/redux/gradientUtils.ts
@@ -1,3 +1,0 @@
-export const calcGradientRange = (range: number) => (range === 2 ? range + 1 : range);
-
-export default {};
